split monolithic AdminController into specific module controllers
