+++ app/app/Providers/RouteServiceProvider.php
@@ -38,7 +38,10 @@ class RouteServiceProvider extends ServiceProvider
                 ->group(base_path('routes/api.php'));
 
             Route::middleware('web')
-                ->group(base_path('routes/web.php'));
+                ->group(base_path('routes/trail.php'));
+                
+            Route::middleware('web')
+                ->group(base_path('routes/dashboard.php'));
         });
     }
