+++ app/routes/trail.php
@@ -1,0 +1,25 @@
+<?php
+
+use Illuminate\Support\Facades\Route;
+
+use App\Http\Controllers\TeamController;
+use App\Http\Controllers\TrailController;
+
+Route::get('/', [TeamController::class, 'index'])->middleware(['checkteam'])->name('start');
+Route::post('/create-team', [TeamController::class, 'create'])->name('create-team');
+Route::get('/join/{id}/{code}', [TeamController::class, 'join'])->middleware(['checkteam'])->name('join-team');
+Route::get('/clone/{id}', [TeamController::class, 'clone'])->middleware(['auth:user']);
+Route::get('/destroy', [TeamController::class, 'destroy'])->middleware(['auth:user', 'auth:team']);
+
+Route::middleware(['auth:team'])->controller(TrailController::class)->group(function() {
+  Route::get('/trail', 'questions')->name('trail');
+  Route::get('/question/{id}', 'viewQuestion')->name('question');
+  Route::post('/question/{id}/submit', 'questionSubmission')->name('submit-question');
+
+  Route::get('/challenges', 'challenges')->name('trail-challenges');
+  Route::get('/challenge/{id}', 'viewChallenge')->name('challenge');
+  Route::post('/challenge/{id}/submit', 'challengeSubmission')->name('submit-challenge');
+  
+  Route::get('/qr', 'showQR')->name('show-qr');
+  Route::get('/qr/code.png', 'getQRImage')->name('qr-image');
+});
